+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.1.1+5
 * Fixed issue where Quiz.multiselect didn't work as expected
+* ReorderList: added [padding] input to set height of initial and last dropzones
 
 # 3.1.1+4
 * Fixed null-safety problem in Quiz